pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "freeclimber"
version = "0.9.0"
description = "Juego de escalada 2D desarrollado para la Junta de Andalucia"
license = "GPL-3.0-or-later"
requires-python = ">=3.9"
dependencies = [
    "pygame-ce>=2.4.0",
]

[project.scripts]
freeclimber = "freeclimber.freeclimber:main"

[tool.setuptools.packages.find]
include = ["freeclimber*"]

[tool.setuptools.package-data]
freeclimber = ["data/**/*"]
